requirements.txt
streamlit
google-genai
pydantic
requests
Pillow